pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "synth-parameter-estimation"
version = "0.1.0"
description = "从音频信号反向估计 Vital 合成器参数"
requires-python = ">=3.10"
dependencies = [
    "pedalboard>=0.9.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "h5py>=3.9.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "soundfile>=0.12.0",
    "muq",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "hypothesis>=6.82.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: 需要外部依赖（Vital VST3、MuQ 模型）的集成测试",
]
